requirements.txt
fastmcp>=2.0.0,<4.0.0
py-synologydsm-api>=2.0.0
aiohttp>=3.9.0
